Introduce Result type into NNAPI
This CL introduces a new error handling type in the NNAPI. This
`Result` object is a discriminated union of either the expected value
or an unexpected error. This CL also adds the convenience macros
NN_ERROR to simplify building and returning an error and NN_TRY to
propagate the error at the call site.

This CL only defines a single templated Result type with the
unexpected state as a std::string to describe the error being
returned. For subsequent CLs, this will be expanded to include
additional error information such as the ErrorStatus being returned
and the OutputShapes in the case of an execution that failed due to
insufficient output buffer size.

Bug: N/A
Test: mma
Change-Id: I9d03e112e3c5489dfa4b585a21abb617a6829f1c
Merged-In: I9d03e112e3c5489dfa4b585a21abb617a6829f1c
(cherry picked from commit 7de00405088452d13ecaf964944a36ad9638d4e7)