model output of unspecified shape as partition input should not force CPU fallback
When execution of one partition writes a model output, and that output's
shape is not fully specified, and the execution succeeds, and that model
output is the input of another partition, execution of that second
partition fails. This results in CPU fallback, except that an OEM or
extension operation results in execution failure being reported at the
NDK level.

The reason execution of the second partition fails is that although
shape information is computed by execution of the first partition, that
shape information is not propagated to the execution of the second
partition: When we try to execute the second partition, both the model
input operand and the request input argument have unspecified shape,
causing a validation failure.

The fix is that in ExecutionStep::mapInputsAndOutputs() when we
initialize the second partition's input from the first partition's
output, we copy the dimensions from the first partition's output to the
second partition's input.

Test: NeuralNetworksTest_static (PartitioningTest, DynamicTemporariesTest, RandomPartitioningTest)

Bug: 168657259
Merged-In: I388f14feb2e55d1632958a07e2f2b5cecb9328d8
Change-Id: I388f14feb2e55d1632958a07e2f2b5cecb9328d8
(cherry picked from commit bf459f7b742bfbb3bc06948014d73d6fd41591ab)